requirements.txt
RPi.GPIO
python-smbus
